Uppdate

+++ dialogabout.cpp
@@ -1,7 +1,7 @@
 #include "dialogabout.h"
 #include "ui_dialogabout.h"
 
-DialogAbout::DialogAbout( float version, QWidget *parent) :
+DialogAbout::DialogAbout( double version, QWidget *parent) :
     QDialog(parent),
     ui(new Ui::DialogAbout)
 {
