+++ source/galaxy/galaxycomponents.h
@@ -6,45 +6,45 @@
 
 class GalaxyComponentBulge : GalaxyComponent {
 public:
-    void componentIntensity(RasterPixel* rp, QVector3D& p, float ival) override;
+    void componentIntensity(RasterPixel* rp, QVector3D& p, double ival) override;
 
-    float calculateIntensity(RasterPixel* rp, QVector3D& p, GalaxyInstance* gi, float weight) override;
-    float getHeightModulation(float height) override;
-    float getRadius(const QVector3D& p, QVector3D& P, float& dott, GalaxyInstance* gi);
+    double calculateIntensity(RasterPixel* rp, QVector3D& p, GalaxyInstance* gi, double weight) override;
+    double getHeightModulation(double height) override;
+    double getRadius(const QVector3D& p, QVector3D& P, double& dott, GalaxyInstance* gi);
 
 };
 
 class GalaxyComponentDisk : GalaxyComponent {
 	
-    void componentIntensity(RasterPixel* rp, QVector3D& p, float ival) override;
+    void componentIntensity(RasterPixel* rp, QVector3D& p, double ival) override;
 };
 
 class GalaxyComponentDust : GalaxyComponent {
 		
-    void componentIntensity(RasterPixel* rp, QVector3D& p, float ival) override;
+    void componentIntensity(RasterPixel* rp, QVector3D& p, double ival) override;
 
 };
 
 class GalaxyComponentStars : GalaxyComponent {
 		
-    void componentIntensity(RasterPixel* rp, QVector3D& r, float ival) override;
+    void componentIntensity(RasterPixel* rp, QVector3D& r, double ival) override;
 		
 };
 
 class GalaxyComponentStarsSmall : GalaxyComponent {
 
-    void componentIntensity(RasterPixel* rp, QVector3D& r, float ival) override;
+    void componentIntensity(RasterPixel* rp, QVector3D& r, double ival) override;
 
 };
 
 class GalaxyComponentDust2 : GalaxyComponent {
 
-    void componentIntensity(RasterPixel* rp, QVector3D& r, float ival) override;
+    void componentIntensity(RasterPixel* rp, QVector3D& r, double ival) override;
 
 };
 
 class GalaxyComponentDustPositive : GalaxyComponent {
 
-    void componentIntensity(RasterPixel* rp, QVector3D& r, float ival) override;
+    void componentIntensity(RasterPixel* rp, QVector3D& r, double ival) override;
 
 };
