+++ source/galaxy/galaxyinstance.h
@@ -8,7 +8,7 @@
 class GalaxyInstance {
 private:
     Galaxy* m_galaxy;
-    float m_intensityScale=1, m_redshift=1;
+    double m_intensityScale=1, m_redshift=1;
     QVector3D m_position = QVector3D(0,0,0);
     QVector3D m_orientation = QVector3D(0,1,0);
     QQuaternion m_rotmat;
@@ -19,20 +19,20 @@ public:
         return m_galaxy;
     }
 
-    float m_workTheta, m_workRad, m_workWinding, m_workLength;
-    float m_winding;
-    float m_currentRadius;
+    double m_workTheta, m_workRad, m_workWinding, m_workLength;
+    double m_winding;
+    double m_currentRadius;
 
     void SetGalaxy(Galaxy* g);
     GalaxyInstance();
     GalaxyInstance* Clone();
     void setupQuaternions();
-    GalaxyInstance(Galaxy* g, QString n, QVector3D p, QVector3D o, float ins, float rs);
+    GalaxyInstance(Galaxy* g, QString n, QVector3D p, QVector3D o, double ins, double rs);
 
-    float intensityScale() const;
-    void setIntensityScale(float intensityScale);
-    float redshift() const;
-    void setRedshift(float redshift);
+    double intensityScale() const;
+    void setIntensityScale(double intensityScale);
+    double redshift() const;
+    void setRedshift(double redshift);
     QQuaternion& rotmat();
     void setRotmat(const QQuaternion &rotmat);
     QVector3D position() const;
