+++ source/galaxy/hpxrasterizer.h
@@ -10,7 +10,7 @@ class HPXRasterizer: public Rasterizer
 {
 
 #ifdef USE_HEALPIX
-    Healpix_Map<float>* m_map = nullptr;//(0,RING);
+    Healpix_Map<double>* m_map = nullptr;//(0,RING);
 #endif
     QQuaternion m_rotMatrix;
     bool m_onlyDust = true;
