+++ source/galaxy/rasterpixel.h
@@ -7,21 +7,21 @@ private:
     QVector3D m_I = QVector3D(0,0,0);
 
 public:
-    float winding, radius;
-    float z;
-    float tmp;
+    double winding, radius;
+    double z;
+    double tmp;
     QVector3D dir;
-    float step;
+    double step;
     QVector3D P;
     QVector3D dust = QVector3D(0,0,0);
 
     void Add(RasterPixel* o);
-    float scale = 1;
+    double scale = 1;
     RasterPixel(RasterPixel* o);
     RasterPixel();
-    void Floor(float f);
+    void Floor(double f);
     QVector3D& I();
     void setI(const QVector3D &I);
-    float getScale() const;
-    void setScale(float value);
+    double getScale() const;
+    void setScale(double value);
 };
